ruby 1.8.7 string slice fix
Ruby 1.8.7 'helpfully' returns the character code instead of the character when
you reference an index in an array. This commit works around that 'helpful' behavior
calling .chr to convert the code to a character, or essentially do nothing in
never versions of ruby.

Also adds ye ol' ruby debugger when you're running 1.8.7.